2020/stopmotion/pyproject.toml
[tool.poetry]
name = "python-video-gen"
version = "0.1.0"
description = ""
authors = ["Nicolai Willems <nicwi@yousee.dk>"]

[tool.poetry.dependencies]
python = "^3.8"
ffmpeg-python = "^0.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
